Made API calls configurable at build time
Added remote and localhost API URLs to environments for production and
dev to remove the need to manually type the base API url. This allows
Google auth to be easily used in dev environment. Show Colors quick
setting is now defaulted to on when first launching app. Fixed a Google
Auth bug where the currentUser was being set as the auth token returned
from the API response, not the actual username